fixed error messages

+++ 3/launch.sh
@@ -23,10 +23,20 @@ if [[ $(cat ~/received_file.txt) == "" ]]; then
         exit 1
 fi
 
-/root/ros_repo/file_equal.py /root/answer_file.txt /root/received_file.txt "not_equal"> error_log
+error_log=$(/root/ros_repo/file_equal.py /root/answer_file.txt /root/received_file.txt "not_equal")
 if [[ $error_log != "" ]]; then
         echo -e "Error: Your Markers are incorrect"
         echo -e "$(cat ~/received_file.txt | grep error)"
+        rm -rf $(cat workspace.txt)/build/ $(cat workspace.txt)/devel/ /root/ros_repo/
+        exit 1
+fi
+
+if [[ $error_log != "" ]] && [[ $(cat ~/received_file.txt | grep error) == "" ]]; then
+        echo -e "Error: Your Markers are incorrect"
+        echo -e "Test data:"
+        echo -e "$(cat ~/answer_file.txt)"
+        echo -e "You provide data:"
+        echo -e "$(cat ~/received_file.txt)"
         rm -rf $(cat workspace.txt)/build/ $(cat workspace.txt)/devel/ /root/ros_repo/
         exit 1
 fi
